backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.52
asyncpg==0.27.0
alembic==1.11.1
pydantic==1.10.11
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
qrcode==7.4
Pillow==10.0.0
aiofiles==23.1.0
pandas==2.1.2
openpyxl==3.1.2
gspread==5.7.2
google-auth==2.20.0
celery==5.3.1
redis==4.6.0
email-validator==2.2.0
python-dotenv==1.0.0
slowapi==0.1.5
